Removing dependencies between tests in DataSourceDefinition test
