+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jpa/datasourcedefinition/persistence.xml
